Prepare CensusSeq for public deploy
Split FindDonorRatiosInBulkData into CensusSeq and Rollcall
Add test coverage
Add executable jars to build.xml
