+++ testdata/org/broadinstitute/dropseq/censusseq/answer_key.txt
@@ -1,0 +1,4 @@
+experiment	sample1	sample2	sample3	sample4	sample5	sample6	sample7	sample8	sample9	sample10	sample11	sample12	sample13	sample14	sample15	sample16	sample17	sample18	sample19	sample20
+knownMixture	0.005	0.015	0.02	0.025	0.035	0.045	0.05	0.08	0.1	0.125	0.005	0.015	0.02	0.025	0.035	0.045	0.05	0.08	0.1	0.125
+iteration	0.0130989824230123	0.016950555687824	0.0233089287434368	0.0266529348565516	0.0365098646708981	0.0490038713721658	0.054208321458227	0.0745277992527529	0.0937681207287736	0.116430316643474	0.0106819799070596	0.0185169653628235	0.0205935654472392	0.0262500953001646	0.0350091473228984	0.0454086152090288	0.0516999962277768	0.077540448843394	0.0924798876255935	0.117359602916905
+optimization	0.0140697912561064	0.0178773203042479	0.0236722699421083	0.0268852547220035	0.0361552771575696	0.0480988224276388	0.0530591212322238	0.0735257940937486	0.0932493274502335	0.117671487460181	0.0116596821738424	0.0193245772714693	0.0212399615778505	0.0264096960863518	0.0346972872260134	0.0444984558123454	0.0505423326564281	0.0768091640110585	0.0921347767966905	0.118419600341888
